+++ cmd/secret_generation.go
@@ -3,23 +3,32 @@ package cmd
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
 	"strings"
+	"sync"
 	"text/template"
 
 	"github.com/Masterminds/sprig/v3"
+	"github.com/helmfile/vals"
 	"go.uber.org/zap"
 	"golang.org/x/crypto/ssh/terminal"
 	"gopkg.in/yaml.v3"
 
+	"rmk/providers/aws_provider"
+	"rmk/providers/azure_provider"
+	"rmk/providers/google_provider"
 	"rmk/util"
 )
 
 // Custom name function for parsing template
 const (
-	Prompt      = "prompt"
-	RequiredEnv = "requiredEnv"
+	FetchSecretValue = "fetchSecretValue"
+	Prompt           = "prompt"
+	RequiredEnv      = "requiredEnv"
+
+	valsCacheSize = 256
 )
 
 type GenerationSpec struct {
@@ -37,6 +46,9 @@ type GenerationRule struct {
 	Name     string `yaml:"name"`
 	Template string `yaml:"template"`
 }
+
+var instance *vals.Runtime
+var once sync.Once
 
 func prompt(name string) (string, error) {
 	fmt.Printf("Enter %s: ", name)
@@ -57,9 +69,61 @@ func requiredEnv(name string) (string, error) {
 	return "", fmt.Errorf("required env var %s is not set", name)
 }
 
+func valsInstance() (*vals.Runtime, error) {
+	var err error
+	once.Do(func() {
+		instance, err = vals.New(vals.Options{CacheSize: valsCacheSize, LogOutput: io.Discard})
+	})
+
+	return instance, err
+}
+
+func fetchSecretValue(path string) (string, error) {
+	valsMap := make(map[string]any)
+	valsMap["key"] = path
+	resultMap, err := expandSecretRefs(valsMap)
+	if err != nil {
+		return "", err
+	}
+
+	rendered, ok := resultMap["key"]
+	if !ok {
+		return "", fmt.Errorf("unexpected error occurred, %v doesn't have 'key' key", resultMap)
+	}
+
+	result, ok := rendered.(string)
+	if !ok {
+		return "", fmt.Errorf("expected %v to be string", rendered)
+	}
+
+	return result, nil
+}
+
+func expandSecretRefs(values map[string]any) (map[string]any, error) {
+	awsEnvs := map[string]string{
+		aws_provider.AWSSDKLoadConfig: "1",
+		aws_provider.AWSSDKGoLogLevel: "off",
+	}
+
+	if err := util.SetOSEnvs(false, awsEnvs); err != nil {
+		return nil, err
+	}
+
+	runtime, err := valsInstance()
+	if err != nil {
+		return nil, err
+	}
+
+	return runtime.Eval(values)
+}
+
 func (gf *GenerationFuncMap) createFuncMap() {
 	gf.funcMap = sprig.TxtFuncMap()
-	for key, val := range map[string]interface{}{RequiredEnv: requiredEnv, Prompt: prompt} {
+	for key, val := range map[string]interface{}{
+		FetchSecretValue: fetchSecretValue,
+		Prompt:           prompt,
+		RequiredEnv:      requiredEnv,
+	} {
 		gf.funcMap[key] = val
 	}
 }
@@ -133,6 +197,31 @@ func (g *GenerationSpec) writeSpecSecrets(force bool) error {
 }
 
 func (sc *SecretCommands) genSpecSecrets(specFiles []string) error {
+	switch sc.Conf.ClusterProvider {
+	case aws_provider.AWSClusterProvider:
+		if sc.Conf.AwsConfigure != nil {
+			if err := sc.Conf.SetAWSCredentialsEnv(false); err != nil {
+				return err
+			}
+		}
+	case azure_provider.AzureClusterProvider:
+		if sc.Conf.AzureConfigure != nil {
+			if err := sc.Conf.AzureConfigure.ReadSPCredentials(sc.Conf.Name); err != nil {
+				return err
+			}
+
+			if err := sc.Conf.SetAzureCredentialsEnv(false); err != nil {
+				return err
+			}
+		}
+	case google_provider.GoogleClusterProvider:
+		if sc.Conf.GCPConfigure != nil {
+			if err := sc.Conf.SetGCPCredentialsEnv(false); err != nil {
+				return err
+			}
+		}
+	}
+
 	genSpec := &GenerationSpec{}
 
 	for _, spec := range specFiles {