+++ docs/release-notes.md
@@ -1,1 +1,4 @@
+- Bumped Go version to v1.24.2.
 - Extended release category to support Helmfile v1.1.X command set.
+- Added fetchSecretValue function for using Vals in secret spec templates.
+- Added cluster provider credentials environment variables for runtime Helmfile.